Load button works
changed the main menu so the data that has been stored by datapersistence can be loaded. Still need to add more functionality to the system overall, such as save points instead of saving on application quit. Also need to be able to make it so if you are in a different scene, that is saved.