Created EnemyMelee, EnemyRanged scripts
Created EnemyMelee, EnemyRanged scripts. Moved code from Character and Enemy into respective enemy scripts. This is because ultimately the 2 don't share that much code and need to be different from top to bottom. Moved a few methods into player from Character as well. Hopefully this makes things a bit easier to find when others are working on the project