Updates to the Data Persistence
updated the saving and loading of the doors, pressure plate, levers. Removed several Debug lines from various scripts. Doors will need updating if we want to have doors that aren't locked, as right now the logic is all doors will be locked until the player unlocks them. Also the box is always starting at a position of 0,0,0. Need to figure out the logic of having it start where we desire.