Player Facing Save
Made it so when the player saves, the direction they were facing is also saved. Did this because it looked funky when saving while the box is held, and the player always loaded facing up, while the box could be to the left. Now the player is facing the same direction as the box.