More DataPersistence work and Spider updates
Put in a check to see if there is any data saved for several items. If there isn't then the attributes saved in the inspector will be used. Previously it would always load false for several items, regardless if we desired them to be true. Changed Box tag to Box, and pressureplate code to only look for that tag. Changed Spider AI to move away after it gets close enough, then resume trying to attack