More Bug Fixes/Organized Hierarchy in each level to be alphabetical/Finished end game scene
1.BUG - Boss would sometimes not begin his teleport phase
FIX - In the SpawnEnemyCo coroutine, the spider and rat enemy instantiate was looking for LintEnemy script, which they did not have. Switched to the proper script
2.BUG - While fighting the boss, you could get knocked into the portal and go back to the previous level.
FIX - Added a line to shut off the portal game object inside the timelinemanager script
3.BUG - If the player attacked an enemy and got hit at nearly the same time, the player could continuously be pushed, and sometimes turn invisible(Null references thrown)
FIX - Added checks in all coroutines everytime a yield waitfor was done to see if the object was null
4.BUG - level_1 door to the left you could walk through even though it should have been closed.
FIX = DoorTakeInput did not have GUID's. added

Organized each scenes hierarchy. Did this for the sake of our professor or whoever has to grade the game. Finally I finished the End Game sequence. You can now restart the game from the ending scene or quit the application