Clearing of warnings and faults. No more warnings
BUG - Null reference every time a scene was loaded for Box and Portal script.
FIX - Found I was making an assignment after the scene was unloaded in the SaveData() method, which threw the fault. Moved assignment into start methods.
BUG - End scene had a null reference to the fade image. but it does not exist in end scene
FIX - Put a check into menucontroller script to see if the current scene is the main menu, if not, then no assignment to the fade image is made
Warning - Lever door in level_2 threw animator not attached to animator controller warning
FIX - unpacked the lever door from the prefab and removed the animator since it wasnt using it