Warning Clearing
1.WARNING - transition needs at least one condition or an exit time to be valid. This was for the crossbowChest
FIX - Added a state for entry that had the sprite simply disappear.

Fixed mistake made after making code for player to be holding crossbow if that was the state the player went to the next scene in. Issue was there was no logic for holding the dagger, so at the start of the game if the player got the dagger, went to the next scene they could not attack. Added a line of code for the trigger of getting the dagger that holding dagger is true.

Finally lowered volume of footsteps.